requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
SQLAlchemy
aiosqlite
asyncpg
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
